Add local image management functions
Moved image listing and uploading functions into a new localImages module. This separation enhances code modularity and improves maintainability by clearly segmenting local image operations.
