+++ lib/imageUploader.ts
@@ -1,13 +1,4 @@
-import fs from "fs"
-import { list, put } from "@vercel/blob"
-
-import {
-  absPathFromPublic,
-  getNonIgnoredFiles,
-  relativePathFromPublic,
-} from "@/lib/fileHelper"
-import { getRedisClient } from "@/lib/utils/redis"
-
+import { put } from "@vercel/blob"
 export async function uploadImageToVercel(
   buffer: Buffer,
   pathFromPublic: string
@@ -33,86 +24,4 @@ function generateImageUrl(pathFromPublic: string) {
     pathFromPublic = pathFromPublic.slice(1)
   }
   return `https://pcpfoetqkuq7jmso.public.blob.vercel-storage.com/${pathFromPublic}`
-}
-
-export async function uploadPublicImagesToVercel() {
-  const urls = []
-  const imageFiles = getImagesFromPublic()
-  for (const pathFromPublic of imageFiles) {
-    let url = await uploadImageToVercelIfNecessary(pathFromPublic)
-    urls.push(url)
-  }
-  return urls
-}
-
-function getImagesFromPublic() {
-  const allFiles = getNonIgnoredFiles(absPathFromPublic(""))
-  const imageFiles = []
-  for (const file of allFiles) {
-    if (
-      file.endsWith(".png") ||
-      file.endsWith(".jpg") ||
-      file.endsWith(".jpeg")
-    ) {
-      imageFiles.push(file)
-    }
-  }
-  return imageFiles.map(relativePathFromPublic)
-}
-
-export async function uploadImageToVercelIfNecessary(pathFromPublic: string) {
-  let url = await vercelImageExists(pathFromPublic)
-  if (url) {
-    console.log(`Image already uploaded: ${pathFromPublic}`)
-    return
-  }
-  const buffer = fs.readFileSync(absPathFromPublic(pathFromPublic))
-  url = await uploadImageToVercel(buffer, pathFromPublic)
-  console.log(`Uploaded: ${url}`)
-  return url
-}
-
-export async function downloadAllBlobImages() {
-  const listBlobResult = await list()
-  const blobs = listBlobResult.blobs
-  // Loop through array of blobs
-  for (const blob of blobs) {
-    const url = blob.url
-    // Download image from url
-    const response = await fetch(url)
-    const buffer = await response.text()
-    let pathname = relativePathFromPublic(blob.pathname)
-    const absPath = absPathFromPublic(pathname)
-    if (fs.existsSync(absPath)) {
-      console.log(`Image already exists: ${absPath}`)
-      continue
-    }
-    fs.writeFileSync(absPath, buffer)
-  }
-}
-
-export async function listImagesOnVercel() {
-  const cacheKey = "vercelImages"
-  const cachedImages = await getRedisClient().get(cacheKey)
-  if (cachedImages) {
-    return JSON.parse(cachedImages)
-  }
-  const listBlobResult = await list()
-  const blobs = listBlobResult.blobs
-  const urls = []
-  for (const blob of blobs) {
-    urls.push(blob.url)
-  }
-  await getRedisClient().set(cacheKey, JSON.stringify(urls))
-  return urls
-}
-
-export async function getVercelImageUrlFromPath(pathFromPublic: string) {
-  const vercelImages = await listImagesOnVercel()
-  for (const url of vercelImages) {
-    if (url.includes(pathFromPublic)) {
-      return url
-    }
-  }
-  return null
 }
